fix(chunk-cache): preserve protocol parity after renderer rewrite
- extract canonical map_chunk serialization and FNV-1a hashing into a pure
  shared module with a fixed parity vector (826a82a5) matching the proxy
- add binary/typed-array/undefined round-trip coverage
- bump minecraft-renderer to 0.1.92, which owns the smart/cave culling
  debug fields introduced by upstream #585
- keep the renderer debug UI lint- and type-safe
