+++ package.json
@@ -51,6 +51,7 @@
     "@nxg-org/mineflayer-auto-jump": "^0.7.18",
     "@nxg-org/mineflayer-tracker": "1.3.0",
     "@react-oauth/google": "^0.12.1",
+    "@royalscome/nextconsole": "^1.0.7",
     "@stylistic/eslint-plugin": "^2.6.1",
     "@types/gapi": "^0.0.47",
     "@types/react": "^18.2.20",
@@ -68,7 +69,6 @@
     "debug": "^4.3.4",
     "deepslate": "^0.23.5",
     "diff-match-patch": "^1.0.5",
-    "@royalscome/nextconsole": "^1.0.7",
     "esbuild": "^0.19.3",
     "esbuild-plugin-polyfill-node": "^0.3.0",
     "express": "^4.18.2",
@@ -84,7 +84,7 @@
     "minecraft-data": "0.0.0",
     "minecraft-inventory": "^0.1.47",
     "minecraft-protocol": "github:PrismarineJS/node-minecraft-protocol#master",
-    "minecraft-renderer": "^0.1.91",
+    "minecraft-renderer": "^0.1.92",
     "mineflayer-item-map-downloader": "github:zardoy/mineflayer-item-map-downloader",
     "mojangson": "^2.0.4",
     "net-browserify": "github:zardoy/prismarinejs-net-browserify",
